+++ src/screen/configScreen.py
@@ -1,4 +1,5 @@
 from time import sleep
+from appContainer import component
 from config.config import Config
 from lib.graphik.src.graphik import Graphik
 from screen.screenType import ScreenType
@@ -6,6 +7,7 @@ from ui.status import Status
 import pygame
 
 # @author Daniel McCoy Stephenson
+@component
 class ConfigScreen:
     def __init__(self, graphik: Graphik, config: Config, status: Status):
         self.graphik = graphik