+++ src/screen/worldScreen.py
@@ -6,6 +6,8 @@ import os
 import time
 import jsonschema
 import pygame
+from appContainer import component
+from di import Container
 from entity.apple import Apple
 from entity.bed import Bed
 from entity.campfire import Campfire
@@ -50,6 +52,7 @@ MIDDLE_MOUSE_BUTTON = 2
 
 # @author Daniel McCoy Stephenson
 # @since August 16th, 2022
+@component
 class WorldScreen:
     def __init__(
         self,
@@ -59,6 +62,7 @@ class WorldScreen:
         tickCounter: TickCounter,
         stats: Stats,
         player: Player,
+        container: Container,
     ):
         self.graphik = graphik
         self.config = config
@@ -66,26 +70,23 @@ class WorldScreen:
         self.tickCounter = tickCounter
         self.stats = stats
         self.player = player
+        self.container = container
         self.running = True
         self.showInventory = False
         self.nextScreen = ScreenType.OPTIONS_SCREEN
         self.changeScreen = False
-        self.roomJsonReaderWriter = RoomJsonReaderWriter(
-            self.config.gridSize, self.graphik, self.tickCounter, self.config
-        )
-        self.mapImageUpdater = MapImageUpdater(self.tickCounter, self.config)
+        self.roomJsonReaderWriter = self.container.resolve(RoomJsonReaderWriter)
+        self.mapImageUpdater = self.container.resolve(MapImageUpdater)
+        self.hudDragManager = self.container.resolve(HudDragManager)
         self.minimapScaleFactor = 0.10
         self.minimapX = 5
         self.minimapY = 5
         self.cursorSlot = InventorySlot()
         self.clock = pygame.time.Clock()
         self.showHelp = False
-        self.hudDragManager = HudDragManager()
 
     def initialize(self):
-        self.map = Map(
-            self.config.gridSize, self.graphik, self.tickCounter, self.config
-        )
+        self.map = self.container.resolve(Map)
 
         # load player location if possible
         if os.path.exists(self.config.pathToSaveDirectory + "/playerLocation.json"):
@@ -116,7 +117,7 @@ class WorldScreen:
         self.initializeLocationWidthAndHeight()
 
         self.status.set("Entered the world")
-        self.energyBar = EnergyBar(self.graphik, self.player)
+        self.energyBar = self.container.resolve(EnergyBar)
 
         self.hudDragManager.register("hotbar", self._getHotbarDefaultRect)
         self.hudDragManager.register("status", lambda: self.status.getDefaultRect())